[git] Change defaultBranch to 'main'
Tbh I liked 'master', not because of the historical origin but because
of the convention. It is used everywhere. Otoh I don't really care.
'main' is shorter. 4 vs 6 character. And if some people go crazy about
it, or companies start measuring offensiveness of the repo, well, better
to stick to 'main' :)